hw1/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
byol-pytorch==0.6.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.5
contourpy==1.1.1
cycler==0.11.0
docker-pycreds==0.4.0
filelock==3.12.4
fonttools==4.42.1
fsspec==2023.9.2
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.37
glob2==0.7
huggingface-hub==0.17.3
idna==3.4
imageio==2.31.5
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
Pillow==10.0.1
protobuf==4.24.4
psutil==5.9.5
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-ranger==0.1.1
pytorch-swats==0.1.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.32.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sympy==1.12
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
timm==0.9.7
torch==2.0.1
torch-optimizer==0.3.0
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
zipp==3.17.0